Move and create testing factory for tests
